+++ Tests/Integration/DSCResources/Firefox.config.ps1
@@ -4,16 +4,38 @@ Configuration Firefox_config
     (
         [Parameter(Mandatory = $true)]
         [string]
-        $StigVersion
+        $StigVersion,
+
+        [Parameter()]
+        [string[]]
+        $SkipRule,
+
+        [Parameter()]
+        [string[]]
+        $SkipRuleType,
+
+        [Parameter()]
+        [string[]]
+        $Exception
     )
 
     Import-DscResource -ModuleName PowerStig
 
     Node localhost
     {
+        & ([scriptblock]::Create("
         Firefox FirefoxConfiguration
         {
-            Stigversion = $StigVersion
-        }
+            Stigversion = '$StigVersion'
+            $(if ($null -ne $Exception)
+            {
+                "Exception = @{'$Exception'= @{'Value'='1234567'}}"
+            })
+            $(if ($null -ne $SkipRule)
+            {
+                "SkipRule = @($( ($SkipRule | % {"'$_'"}) -join ',' ))`n"
+            })
+        }")
+        )
     }
 }