+++ Tests/Integration/DSCResources/IisServer.config.ps1
@@ -12,17 +12,43 @@ Configuration IisServer_Config
 
         [Parameter(Mandatory = $true)]
         [string]
-        $LogPath
+        $LogPath,
+
+        [Parameter()]
+        [string[]]
+        $SkipRule,
+
+        [Parameter()]
+        [string[]]
+        $SkipRuleType,
+
+        [Parameter()]
+        [string[]]
+        $Exception
     )
 
     Import-DscResource -ModuleName PowerStig
     Node localhost
     {
+        & ([scriptblock]::Create("
         IisServer ServerConfiguration
         {
-            OsVersion   = $OsVersion
-            StigVersion = $StigVersion
-            LogPath     = $LogPath
-        }
+            OsVersion = '$OsVersion'
+            StigVersion = '$StigVersion'
+            LogPath = '$LogPath'
+            $(if ($null -ne $Exception)
+            {
+                "Exception = @{'$Exception'= @{'Value'='1234567'}}"
+            })
+            $(if ($null -ne $SkipRule)
+            {
+                "SkipRule = @($( ($SkipRule | % {"'$_'"}) -join ',' ))`n"
+            }
+            if ($null -ne $SkipRuleType)
+            {
+                "SkipRuleType = @($( ($SkipRuleType | % {"'$_'"}) -join ',' ))`n"
+            })
+        }")
+        )
     }
 }