+++ Tests/Integration/DSCResources/IisServer.integration.tests.ps1
@@ -14,6 +14,8 @@ try
     #region Integration Tests
     foreach ($stig in $stigList)
     {
+        [xml] $dscXml = Get-Content -Path $stig.Path
+
         Describe "IIS Server $($stig.StigVersion) mof output" {
 
             It 'Should compile the MOF without throwing' {
@@ -23,10 +25,8 @@ try
                         -StigVersion $stig.StigVersion `
                         -LogPath $TestDrive `
                         -OutputPath $TestDrive
-                } | Should not throw
+                } | Should -Not -Throw
             }
-
-            [xml] $dscXml = Get-Content -Path $stig.Path
 
             $configurationDocumentPath = "$TestDrive\localhost.mof"
 
@@ -39,34 +39,34 @@ try
 
                 foreach ($setting in $dscXml)
                 {
-                    if (-not ($dscMof.ResourceID -match $setting.Id) )
+                    if (-not ($dscMof.ResourceID -match $setting.id) )
                     {
-                        Write-Warning -Message "WebServer missing IisLoggingRule Setting $($setting.Id)"
+                        Write-Warning -Message "WebServer missing IisLoggingRule Setting $($setting.id)"
                         $hasAllSettings = $false
                     }
                 }
 
                 It "Should have $($dscXml.count) IisLoggingRule settings" {
-                    $hasAllSettings | Should Be $true
+                    $hasAllSettings | Should -Be $true
                 }
             }
 
             Context 'WebConfigurationPropertyRule' {
                 $hasAllSettings = $true
                 $dscXml = $dscXml.DISASTIG.WebConfigurationPropertyRule.Rule
-                $dscMof = $instances | Where-Object {$PSItem.ResourceID -match "\[xWebConfigProperty\]"}
+                $dscMof = $instances | Where-Object -FilterScript {$PSItem.ResourceID -match "\[xWebConfigProperty\]"}
 
                 foreach ($setting in $dscXml)
                 {
-                    if (-not ($dscMof.ResourceID -match $setting.Id) )
+                    if (-not ($dscMof.ResourceID -match $setting.id) )
                     {
-                        Write-Warning -Message "WebServer missing WebConfigurationPropertyRule Setting $($setting.Id)"
+                        Write-Warning -Message "WebServer missing WebConfigurationPropertyRule Setting $($setting.id)"
                         $hasAllSettings = $false
                     }
                 }
 
 
-                It "Should have $($dscXml.Count) WebConfigurationPropertyRule settings" {
+                It "Should have $($dscXml.count) WebConfigurationPropertyRule settings" {
                     $hasAllSettings | Should -Be $true
                 }
             }
@@ -74,20 +74,108 @@ try
             Context 'MimeTypeRule' {
                 $hasAllSettings = $true
                 $dscXml = $dscXml.DISASTIG.MimeTypeRule.Rule
-                $dscMof = $instances | Where-Object {$PSItem.ResourceID -match "\[xIisMimeTypeMapping\]"}
+                $dscMof = $instances | Where-Object -FilterScript {$PSItem.ResourceID -match "\[xIisMimeTypeMapping\]"}
 
                 foreach ($setting in $dscXml)
                 {
-                    if (-not ($dscMof.ResourceID -match $setting.Id) )
+                    if (-not ($dscMof.ResourceID -match $setting.id) )
                     {
-                        Write-Warning -Message "WebServer missing MimeTypeRule Setting $($setting.Id)"
+                        Write-Warning -Message "WebServer missing MimeTypeRule Setting $($setting.id)"
                         $hasAllSettings = $false
                     }
                 }
 
-                It "Should have $($dscXml.Count) MimeTypeRule settings" {
-                    $hasAllSettings | Should Be $true
+                It "Should have $($dscXml.count) MimeTypeRule settings" {
+                    $hasAllSettings | Should -Be $true
                 }
+            }
+        }
+
+        Describe "IIS Server $($stig.StigVersion) Single SkipRule/RuleType mof output" {
+
+            $skipRule = Get-Random -InputObject $dscXml.DISASTIG.MimeTypeRule.Rule.id
+            $skipRuleType = "IisLoggingRule"
+
+            It 'Should compile the MOF without throwing' {
+                {
+                    & "$($script:DSCCompositeResourceName)_config" `
+                        -OsVersion $stig.TechnologyVersion `
+                        -StigVersion $stig.StigVersion `
+                        -LogPath $TestDrive `
+                        -OutputPath $TestDrive `
+                        -SkipRule $skipRule `
+                        -SkipRuleType $skipRuleType `
+                } | Should -Not -Throw
+            }
+
+            #region Gets the mof content
+            $configurationDocumentPath = "$TestDrive\localhost.mof"
+            $instances = [Microsoft.PowerShell.DesiredStateConfiguration.Internal.DscClassCache]::ImportInstances($configurationDocumentPath, 4)
+            #endregion
+
+            Context 'Skip check' {
+                #region counts how many Skips there are and how many there should be.
+                $dscIisLoggingRuleXml = $dscXml.DISASTIG.IisLoggingRule.Rule | Where-Object -FilterScript {$_.ConversionStatus -eq "pass"}
+                $expectedSkipRuleCount = ($($dscIisLoggingRuleXml.count) + $($skipRule.count))
+                $dscMof = $instances | Where-Object -FilterScript {$PSItem.ResourceID -match "\[Skip\]"}
+                #endregion
+                It "Should have $expectedSkipRuleCount Skipped settings" {
+                    $dscMof.count | Should -Be $expectedSkipRuleCount
+                }
+            }
+        }
+
+        Describe "IIS Server $($stig.StigVersion) Multiple SkipRule/RuleType mof output" {
+
+            $skipRule = Get-Random -InputObject $dscXml.DISASTIG.WebConfigurationPropertyRule.Rule.id -Count 2
+            $skipRuleType = @('MimeTypeRule','IisLoggingRule')
+
+            It 'Should compile the MOF without throwing' {
+                {
+                    & "$($script:DSCCompositeResourceName)_config" `
+                        -OsVersion $stig.TechnologyVersion `
+                        -StigVersion $stig.StigVersion `
+                        -LogPath $TestDrive `
+                        -OutputPath $TestDrive `
+                        -SkipRule $skipRule `
+                        -SkipRuleType $skipRuleType `
+                } | Should -Not -Throw
+            }
+
+            #region Gets the mof content
+            $configurationDocumentPath = "$TestDrive\localhost.mof"
+            $instances = [Microsoft.PowerShell.DesiredStateConfiguration.Internal.DscClassCache]::ImportInstances($configurationDocumentPath, 4)
+            #endregion
+
+            Context 'Skip check' {
+
+                #region counts how many Skips there are and how many there should be.
+                $dscMimeTypeRuleXml = $dscXml.DISASTIG.MimeTypeRule.Rule | Where-Object -FilterScript {$_.ConversionStatus -eq "Pass"}
+                $dscIisLoggingRuleXml = $dscXml.DISASTIG.IisLoggingRule.Rule | Where-Object -FilterScript {$_.ConversionStatus -eq "Pass"}
+                $expectedSkipRuleCount = ($($dscMimeTypeRuleXml.count) + $($dscIisLoggingRuleXml.count) + $($skipRule.count))
+                $dscMof = $instances | Where-Object -FilterScript {$PSItem.ResourceID -match "\[Skip\]"}
+                #endregion
+
+                It "Should have $expectedSkipRuleCount Skipped settings" {
+                    $dscMof.count | Should -Be $expectedSkipRuleCount
+                }
+            }
+        }
+
+        Describe "IIS Server $($stig.StigVersion) Exception mof output" {
+
+            $exceptionRule = Get-Random -InputObject $dscXml.DISASTIG.WebConfigurationPropertyRule.Rule
+            $exception = $exceptionRule.id
+
+            It "Should compile the MOF with STIG exception $exception without throwing" {
+                {
+                    & "$($script:DSCCompositeResourceName)_config" `
+                        -OsVersion $stig.TechnologyVersion `
+                        -StigVersion $stig.StigVersion `
+                        -LogPath $TestDrive `
+                        -OutputPath $TestDrive `
+                        -Exception $exception
+                } | Should -Not -Throw
             }
         }
     }