+++ Tests/Integration/DSCResources/SqlServer.integration.tests.ps1
@@ -16,12 +16,13 @@ try
     {
         [xml] $dscXml = Get-Content -Path $stig.Path
 
-        Describe "Browser $($stig.TechnologyRole) $($stig.StigVersion) mof output" {
+        Describe "SqlServer $($stig.TechnologyRole) $($stig.StigVersion) mof output" {
 
             It 'Should compile the MOF without throwing' {
                 {
-                    & "$($script:DSCCompositeResourceName)_config" `
-                        -BrowserVersion $stig.TechnologyRole `
+                    & "$($script:DSCCompositeResourceName)$($stig.TechnologyRole)_config" `
+                        -SqlVersion $stig.TechnologyVersion `
+                        -SqlRole $stig.TechnologyRole`
                         -StigVersion $stig.StigVersion `
                         -OutputPath $TestDrive
                 } | Should -Not -Throw
@@ -30,37 +31,40 @@ try
             $configurationDocumentPath = "$TestDrive\localhost.mof"
             $instances = [Microsoft.PowerShell.DesiredStateConfiguration.Internal.DscClassCache]::ImportInstances($configurationDocumentPath, 4)
 
-            Context 'Registry' {
+            Context 'SqlScriptQuery' {
                 $hasAllSettings = $true
-                $dscXml = $dscXml.DISASTIG.RegistryRule.Rule
+                $dscXml = $dscXml.DISASTIG.SqlScriptQueryRule.Rule
                 $dscMof = $instances |
-                    Where-Object {$PSItem.ResourceID -match "\[xRegistry\]" -or $PSItem.ResourceID -match "\[cAdministrativeTemplateSetting\]"}
+                    Where-Object {$PSItem.ResourceID -match "\[SqlScriptQuery\]"}
 
                 foreach ( $setting in $dscXml )
                 {
                     If (-not ($dscMof.ResourceID -match $setting.id) )
                     {
-                        Write-Warning -Message "Missing registry Setting $($setting.id)"
+                        Write-Warning -Message "Missing SqlScriptQuery $($setting.id)"
                         $hasAllSettings = $false
                     }
                 }
 
-                It "Should have $($dscXml.count) Registry settings" {
+                It "Should have $($dscXml.id.count) SqlScriptQueryRule settings" {
                     $hasAllSettings | Should -Be $true
                 }
             }
         }
 
-        Describe "Browser $($stig.TechnologyRole) $($stig.StigVersion) Single SkipRule mof output" {
+        Describe "SqlServer $($stig.TechnologyRole) $($stig.TechnologyVersion) $($stig.StigVersion) Single SkipRule/RuleType mof output" {
 
-            $skipRule = Get-Random -InputObject $dscXml.DISASTIG.RegistryRule.Rule.id
+            $skipRule = Get-Random -InputObject $dscXml.DISASTIG.SqlScriptQueryRule.Rule.id
+            $skipRuleType = "DocumentRule"
 
             It 'Should compile the MOF without throwing' {
                 {
-                    & "$($script:DSCCompositeResourceName)_config" `
-                        -BrowserVersion $stig.TechnologyRole `
+                    & "$($script:DSCCompositeResourceName)$($stig.TechnologyRole)_config" `
+                        -SqlVersion $stig.TechnologyVersion `
+                        -SqlRole $stig.TechnologyRole`
                         -StigVersion $stig.StigVersion `
                         -SkipRule $skipRule `
+                        -SkipRuleType $skipRuleType `
                         -OutputPath $TestDrive
                 } | Should -Not -Throw
             }
@@ -73,8 +77,9 @@ try
             Context 'Skip check' {
 
                 #region counts how many Skips there are and how many there should be.
-                $dscXml = $skipRule.count
-                $dscMof = @($instances | Where-Object -FilterScript {$PSItem.ResourceID -match "\[Skip\]"})
+                $dscXml = $dscXml.DISASTIG.DocumentRule.Rule | Where-Object {$_.ConversionStatus -eq 'pass'}
+                $dscXml = $dscXml.count + $skipRule.count
+                $dscMof = $instances | Where-Object {$PSItem.ResourceID -match "\[Skip\]"}
                 #endregion
 
                 It "Should have $dscXml Skipped settings" {
@@ -83,16 +88,19 @@ try
             }
         }
 
-        Describe "Browser $($stig.TechnologyRole) $($stig.StigVersion) Multiple SkipRule mof output" {
+        Describe "SqlServer $($stig.TechnologyRole) $($stig.TechnologyVersion) $($stig.StigVersion) Multiple SkipRule/RuleType mof output" {
 
-            $skipRule = Get-Random -InputObject $dscXml.DISASTIG.RegistryRule.Rule.id -Count 2
+            $skipRule = Get-Random -InputObject $dscXml.DISASTIG.SqlScriptQueryRule.Rule.id -Count 2
+            $skipRuleType = "DocumentRule"
 
             It 'Should compile the MOF without throwing' {
                 {
-                    & "$($script:DSCCompositeResourceName)_config" `
-                        -BrowserVersion $stig.TechnologyRole `
+                    & "$($script:DSCCompositeResourceName)$($stig.TechnologyRole)_config" `
+                        -SqlVersion $stig.TechnologyVersion `
+                        -SqlRole $stig.TechnologyRole`
                         -StigVersion $stig.StigVersion `
                         -SkipRule $skipRule `
+                        -SkipRuleType $skipRuleType `
                         -OutputPath $TestDrive
                 } | Should -Not -Throw
             }
@@ -105,7 +113,8 @@ try
             Context 'Skip check' {
 
                 #region counts how many Skips there are and how many there should be.
-                $expectedSkipRuleCount = $skipRule.count
+                $dscDocumentTypeRuleXml = $dscXml.DISASTIG.DocumentRule.Rule | Where-Object {$_.ConversionStatus -eq 'pass'}
+                $expectedSkipRuleCount = $dscDocumentTypeRuleXml.count + $skipRule.count
                 $dscMof = $instances | Where-Object -FilterScript {$PSItem.ResourceID -match "\[Skip\]"}
                 #endregion
 
@@ -115,15 +124,16 @@ try
             }
         }
 
-        Describe "Browser $($stig.TechnologyRole) $($stig.StigVersion) Exception mof output" {
+        Describe "SqlServer $($stig.TechnologyRole) $($stig.TechnologyVersion) $($stig.StigVersion) Exception mof output" {
 
-            $exceptionRule = Get-Random -InputObject $dscXml.DISASTIG.RegistryRule.Rule
+            $exceptionRule = Get-Random -InputObject $dscXml.DISASTIG.SqlScriptQueryRule.rule
             $exception = $exceptionRule.id
 
             It "Should compile the MOF with STIG exception $exception without throwing" {
                 {
-                    & "$($script:DSCCompositeResourceName)_config" `
-                        -BrowserVersion $stig.TechnologyRole `
+                    & "$($script:DSCCompositeResourceName)$($stig.TechnologyRole)_config" `
+                        -SqlVersion $stig.TechnologyVersion `
+                        -SqlRole $stig.TechnologyRole`
                         -StigVersion $stig.StigVersion `
                         -OutputPath $TestDrive `
                         -Exception $exception