+++ Tests/Integration/DSCResources/WindowsServer.integration.tests.ps1
@@ -27,7 +27,7 @@ try
                         -ForestName 'integration.test' `
                         -DomainName 'integration.test' `
                         -OutputPath $TestDrive
-                } | Should not throw
+                } | Should -Not -Throw
             }
 
             $configurationDocumentPath = "$TestDrive\localhost.mof"
@@ -36,21 +36,21 @@ try
 
             Context 'AuditPolicy' {
                 $hasAllSettings = $true
-                $dscXml         = $dscXml.DISASTIG.AuditPolicyRule.Rule
-                $dscMof         = $instances |
+                $dscXml = $dscXml.DISASTIG.AuditPolicyRule.Rule
+                $dscMof = $instances |
                 Where-Object {$PSItem.ResourceID -match "\[AuditPolicySubcategory\]"}
 
                 foreach ( $setting in $dscXml )
                 {
-                    If (-not ($dscMof.ResourceID -match $setting.Id) )
+                    If (-not ($dscMof.ResourceID -match $setting.id) )
                     {
-                        Write-Warning -Message "Missing Audit Policy Setting $($setting.Id)"
+                        Write-Warning -Message "Missing Audit Policy Setting $($setting.id)"
                         $hasAllSettings = $false
                     }
                 }
 
-                It "Should have $($dscXml.Count) Audit Policy settings" {
-                    $hasAllSettings | Should Be $true
+                It "Should have $($dscXml.count) Audit Policy settings" {
+                    $hasAllSettings | Should -Be $true
                 }
             }
 
@@ -63,62 +63,62 @@ try
                     following where cmdlet
                 #>
                 $dscXmlPermissionPolicy = $dscXml.DISASTIG.PermissionRule.Rule |
-                    Where-Object { $PSItem.conversionstatus -eq "pass" -and
+                    Where-Object { $PSItem.conversionstatus -eq 'pass' -and
                                    $PSItem.dscResource -ne "ActiveDirectoryAuditRuleEntry"}
                 $dscMofPermissionPolicy = $instances |
                     Where-Object {$PSItem.ResourceID -match "\[NTFSAccessEntry\]|\[RegistryAccessEntry\]"}
 
                 foreach ($setting in $dscXmlPermissionPolicy)
                 {
-                    If (-not ($dscMofPermissionPolicy.ResourceID -match $setting.Id) )
+                    If (-not ($dscMofPermissionPolicy.ResourceID -match $setting.id) )
                     {
-                        Write-Warning -Message "Missing permission setting $($setting.Id)"
+                        Write-Warning -Message "Missing permission setting $($setting.id)"
                         $hasAllSettings = $false
                     }
                 }
 
-                It "Should have $($dscXmlPermissionPolicy.Count) permission settings" {
-                    $hasAllSettings | Should Be $true
+                It "Should have $($dscXmlPermissionPolicy.count) permission settings" {
+                    $hasAllSettings | Should -Be $true
                 }
             }
 
             Context 'Registry' {
                 $hasAllSettings = $true
-                $dscXml   = $dscXml.DISASTIG.RegistryRule.Rule
-                $dscMof   = $instances |
+                $dscXml = $dscXml.DISASTIG.RegistryRule.Rule
+                $dscMof = $instances |
                     Where-Object {$PSItem.ResourceID -match "\[xRegistry\]" -or $PSItem.ResourceID -match "\[cAdministrativeTemplateSetting\]"}
 
                 foreach ( $setting in $dscXml )
                 {
-                    If (-not ($dscMof.ResourceID -match $setting.Id) )
+                    If (-not ($dscMof.ResourceID -match $setting.id) )
                     {
-                        Write-Warning -Message "Missing registry Setting $($setting.Id)"
+                        Write-Warning -Message "Missing registry Setting $($setting.id)"
                         $hasAllSettings = $false
                     }
                 }
 
-                It "Should have $($dscXml.Count) Registry settings" {
-                    $hasAllSettings | Should Be $true
+                It "Should have $($dscXml.count) Registry settings" {
+                    $hasAllSettings | Should -Be $true
                 }
             }
 
             Context 'WMI' {
                 $hasAllSettings = $true
-                $dscXml    = $dscXml.DISASTIG.WmiRule.Rule
-                $dscMof   = $instances |
+                $dscXml = $dscXml.DISASTIG.WmiRule.Rule
+                $dscMof = $instances |
                     Where-Object {$PSItem.ResourceID -match "\[script\]"}
 
                 foreach ( $setting in $dscXml )
                 {
-                    If (-not ($dscMof.ResourceID -match $setting.Id) )
+                    If (-not ($dscMof.ResourceID -match $setting.id) )
                     {
-                        Write-Warning -Message "Missing wmi setting $($setting.Id)"
+                        Write-Warning -Message "Missing wmi setting $($setting.id)"
                         $hasAllSettings = $false
                     }
                 }
 
-                It "Should have $($dscXml.Count) wmi settings" {
-                    $hasAllSettings | Should Be $true
+                It "Should have $($dscXml.count) wmi settings" {
+                    $hasAllSettings | Should -Be $true
                 }
             }
 
@@ -130,15 +130,15 @@ try
 
                 foreach ( $setting in $dscXml )
                 {
-                    If (-not ($dscMof.ResourceID -match $setting.Id) )
+                    If (-not ($dscMof.ResourceID -match $setting.id) )
                     {
-                        Write-Warning -Message "Missing service setting $($setting.Id)"
+                        Write-Warning -Message "Missing service setting $($setting.id)"
                         $hasAllSettings = $false
                     }
                 }
 
-                It "Should have $($dscXml.Count) service settings" {
-                    $hasAllSettings | Should Be $true
+                It "Should have $($dscXml.count) service settings" {
+                    $hasAllSettings | Should -Be $true
                 }
             }
 
@@ -150,15 +150,15 @@ try
 
                 foreach ( $setting in $dscXml )
                 {
-                    If (-not ($dscMof.ResourceID -match $setting.Id) )
+                    If (-not ($dscMof.ResourceID -match $setting.id) )
                     {
-                        Write-Warning -Message "Missing security setting $($setting.Id)"
+                        Write-Warning -Message "Missing security setting $($setting.id)"
                         $hasAllSettings = $false
                     }
                 }
 
-                It "Should have $($dscXml.Count) security settings" {
-                    $hasAllSettings | Should Be $true
+                It "Should have $($dscXml.count) security settings" {
+                    $hasAllSettings | Should -Be $true
                 }
             }
 
@@ -170,15 +170,15 @@ try
 
                 foreach ( $setting in $dscXml )
                 {
-                    If (-not ($dscMof.ResourceID -match $setting.Id) )
+                    If (-not ($dscMof.ResourceID -match $setting.id) )
                     {
-                        Write-Warning -Message "Missing user right $($setting.Id)"
+                        Write-Warning -Message "Missing user right $($setting.id)"
                         $hasAllSettings = $false
                     }
                 }
 
-                It "Should have $($dscXml.Count) user rights settings" {
-                    $hasAllSettings | Should Be $true
+                It "Should have $($dscXml.count) user rights settings" {
+                    $hasAllSettings | Should -Be $true
                 }
             }
 
@@ -190,15 +190,15 @@ try
 
                 foreach ( $setting in $dscXml )
                 {
-                    If (-not ($dscMof.ResourceID -match $setting.Id) )
+                    If (-not ($dscMof.ResourceID -match $setting.id) )
                     {
-                        Write-Warning -Message "Missing security setting $($setting.Id)"
+                        Write-Warning -Message "Missing security setting $($setting.id)"
                         $hasAllSettings = $false
                     }
                 }
 
-                It "Should have $($dscXml.Count) security settings" {
-                    $hasAllSettings | Should Be $true
+                It "Should have $($dscXml.count) security settings" {
+                    $hasAllSettings | Should -Be $true
                 }
             }
 
@@ -210,36 +210,36 @@ try
 
                 foreach ($setting in $dscXml)
                 {
-                    If (-not ($dscMof.ResourceID -match $setting.Id) )
+                    If (-not ($dscMof.ResourceID -match $setting.id) )
                     {
-                        Write-Warning -Message "Missing windows feature $($setting.Id)"
+                        Write-Warning -Message "Missing windows feature $($setting.id)"
                         $hasAllSettings = $false
                     }
                 }
 
-                It "Should have $($dscXml.Count) windows feature settings" {
-                    $hasAllSettings | Should Be $true
+                It "Should have $($dscXml.count) windows feature settings" {
+                    $hasAllSettings | Should -Be $true
                 }
             }
         }
 
         Describe "Windows $($stig.TechnologyVersion) $($stig.TechnologyRole) $($stig.StigVersion) Single SkipRule/RuleType mof output" {
 
-            $SkipRule     = Get-Random -InputObject $dscXml.DISASTIG.RegistryRule.Rule.id
-            $SkipRuleType = "AuditPolicyRule"
+            $skipRule = Get-Random -InputObject $dscXml.DISASTIG.RegistryRule.Rule.id
+            $skipRuleType = "AuditPolicyRule"
 
             It 'Should compile the MOF without throwing' {
                 {
                     & "$($script:DSCCompositeResourceName)_config" `
-                        -OsVersion $stig.TechnologyVersion  `
+                        -OsVersion $stig.TechnologyVersion `
                         -OsRole $stig.TechnologyRole `
                         -StigVersion $stig.StigVersion `
                         -ForestName 'integration.test' `
                         -DomainName 'integration.test' `
-                        -SkipRule $SkipRule `
-                        -SkipRuleType $SkipRuleType `
+                        -SkipRule $skipRule `
+                        -SkipRuleType $skipRuleType `
                         -OutputPath $TestDrive
-                } | Should not throw
+                } | Should -Not -Throw
             }
 
             #region Gets the mof content
@@ -250,35 +250,35 @@ try
             Context 'Skip check' {
 
                 #region counts how many Skips there are and how many there should be.
-                $dscXml = $dscXml.DISASTIG.AuditPolicyRule.Rule | Where-Object {$_.ConversionStatus -eq "pass"}
-                $dscXml = ($($dscXml.Count) + $($SkipRule.Count))
+                $dscXml = $dscXml.DISASTIG.AuditPolicyRule.Rule | Where-Object {$_.ConversionStatus -eq 'pass'}
+                $dscXml = $dscXml.count + $skipRule.count
 
                 $dscMof = $instances | Where-Object {$PSItem.ResourceID -match "\[Skip\]"}
                 #endregion
 
                 It "Should have $dscXml Skipped settings" {
-                    $dscMof.count | Should Be $dscXml
+                    $dscMof.count | Should -Be $dscXml
                 }
             }
         }
 
         Describe "Windows $($stig.TechnologyVersion) $($stig.TechnologyRole) $($stig.StigVersion) Multiple SkipRule/RuleType mof output" {
 
-            $SkipRule     = Get-Random -InputObject $dscXml.DISASTIG.RegistryRule.Rule.id -Count 2
-            $SkipRuleType = @('AuditPolicyRule','AccountPolicyRule')
+            $skipRule = Get-Random -InputObject $dscXml.DISASTIG.RegistryRule.Rule.id -Count 2
+            $skipRuleType = @('AuditPolicyRule','AccountPolicyRule')
 
             It 'Should compile the MOF without throwing' {
                 {
                     & "$($script:DSCCompositeResourceName)_config" `
-                        -OsVersion $stig.TechnologyVersion  `
+                        -OsVersion $stig.TechnologyVersion `
                         -OsRole $stig.TechnologyRole `
                         -StigVersion $stig.StigVersion `
                         -ForestName 'integration.test' `
                         -DomainName 'integration.test' `
-                        -SkipRule $SkipRule `
-                        -SkipRuleType $SkipRuleType `
+                        -SkipRule $skipRule `
+                        -SkipRuleType $skipRuleType `
                         -OutputPath $TestDrive
-                } | Should not throw
+                } | Should -Not -Throw
             }
 
             #region Gets the mof content
@@ -292,14 +292,31 @@ try
                 $dscAuditXml = $dscXml.DISASTIG.AuditPolicyRule.Rule | Where-Object {$_.ConversionStatus -eq "Pass"}
                 $dscPermissionXml = $dscXml.DISASTIG.AccountPolicyRule.Rule | Where-Object {$_.ConversionStatus -eq "Pass"}
 
-                $dscXml = ($($dscAuditXml.Count) + $($dscPermissionXml.count) + $($SkipRule.Count))
-
+                $dscXml = $dscAuditXml.count + $dscPermissionXml.count + $skipRule.count
                 $dscMof = $instances | Where-Object {$PSItem.ResourceID -match "\[Skip\]"}
                 #endregion
 
                 It "Should have $dscXml Skipped settings" {
-                    $dscMof.count | Should Be $dscXml
+                    $dscMof.count | Should -Be $dscXml
                 }
+            }
+        }
+
+        Describe "Windows $($stig.TechnologyVersion) $($stig.TechnologyRole) $($stig.StigVersion) Exception mof output" {
+
+            $exceptionRule = Get-Random -InputObject $dscXml.DISASTIG.RegistryRule.Rule
+            $exception = $exceptionRule.id
+
+            It "Should compile the MOF with STIG exception $exception without throwing" {
+                {
+                    & "$($script:DSCCompositeResourceName)_config" `
+                        -OsVersion $stig.TechnologyVersion `
+                        -OsRole $stig.TechnologyRole `
+                        -StigVersion $stig.StigVersion `
+                        -ForestName 'integration.test' `
+                        -DomainName 'integration.test' `
+                        -Exception $exception
+                } | Should -Not -Throw
             }
         }
     }